add: simple impl of build/search of hnsw

+++ src/search/mod.rs
@@ -48,7 +48,7 @@ pub fn knn_search(using_index: IndexType, query: Rc<VecData>, k_for_search: usiz
         IndexType::HNSW => {
             println!("[Info] Use HNSW index (graph-based)");
             let pg_max_degree: index::hnsw::Degree = 24;
-            let num_layers: usize = 1;
+            let num_layers: u8 = 1;
 
             let timer = time::Instant::now();
             let mut index = index::hnsw::Index::build(dataset, pg_max_degree, num_layers);
